requirements.txt
aiogram==3.13.0
requests==2.32.3
pydantic-settings==2.3.4
pydantic==2.9.1
python-dotenv~=1.0.1